requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
sentence-transformers==2.2.2
torch>=2.2.0